# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumo Task API",
    "version": "1.0.0",
    "description": "REST API for Lumo Task — an Eisenhower-matrix task manager.\n\n**This document is generated** from the route registry in `@lumo/contracts` (`src/registry.ts`) and the Zod schemas it references. It is never hand-edited, and a parity test fails the build if it stops matching the routes the server actually mounts.\n\n**Authentication:** every `/v1/*` route requires a Bearer token except the auth and OAuth handshake endpoints. Call `POST /v1/auth/signin` to obtain one, then click **Authorize** and paste it.\n\n**Errors** all share one envelope: `{ error: { code, message, fields? } }`."
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string"
          },
          "accent": {
            "type": "string"
          },
          "density": {
            "type": "string"
          },
          "reduced_motion": {
            "type": "boolean"
          },
          "ai_enabled": {
            "type": "boolean"
          },
          "pomodoro_duration": {
            "type": "number"
          },
          "short_break": {
            "type": "number"
          },
          "long_break": {
            "type": "number"
          },
          "long_break_interval": {
            "type": "number"
          },
          "auto_start_breaks": {
            "type": "boolean"
          },
          "notifications_enabled": {
            "type": "boolean"
          },
          "morning_reminder_time": {
            "type": "string"
          },
          "evening_reminder_time": {
            "type": "string"
          },
          "due_alerts_enabled": {
            "type": "boolean"
          },
          "onboarding_complete": {
            "type": "boolean"
          },
          "ai_provider": {
            "type": "string"
          },
          "ai_provider_configs": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "hasKey": {
                  "type": "boolean"
                },
                "model": {
                  "type": "string"
                },
                "baseUrl": {
                  "type": "string"
                }
              },
              "required": [
                "hasKey",
                "model",
                "baseUrl"
              ]
            }
          },
          "ai_cloud_enabled": {
            "type": "boolean"
          },
          "ai_cloud_used": {
            "type": "number"
          },
          "ai_cloud_limit": {
            "type": "number"
          }
        },
        "required": [
          "locale",
          "accent",
          "density",
          "reduced_motion",
          "ai_enabled",
          "pomodoro_duration",
          "short_break",
          "long_break",
          "long_break_interval",
          "auto_start_breaks",
          "notifications_enabled",
          "morning_reminder_time",
          "evening_reminder_time",
          "due_alerts_enabled",
          "onboarding_complete",
          "ai_provider",
          "ai_provider_configs",
          "ai_cloud_enabled",
          "ai_cloud_used",
          "ai_cloud_limit"
        ]
      },
      "SettingsPatchBody": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "zh"
            ]
          },
          "accent": {
            "type": "string",
            "enum": [
              "green",
              "cyan",
              "amber",
              "graphite"
            ]
          },
          "density": {
            "type": "string",
            "enum": [
              "comfortable",
              "compact"
            ]
          },
          "reduced_motion": {
            "type": "boolean"
          },
          "ai_enabled": {
            "type": "boolean"
          },
          "pomodoro_duration": {
            "type": "integer"
          },
          "short_break": {
            "type": "integer"
          },
          "long_break": {
            "type": "integer"
          },
          "long_break_interval": {
            "type": "integer"
          },
          "auto_start_breaks": {
            "type": "boolean"
          },
          "notifications_enabled": {
            "type": "boolean"
          },
          "morning_reminder_time": {
            "type": "string",
            "nullable": true
          },
          "evening_reminder_time": {
            "type": "string",
            "nullable": true
          },
          "due_alerts_enabled": {
            "type": "boolean"
          },
          "onboarding_complete": {
            "type": "boolean"
          },
          "ai_provider": {
            "type": "string",
            "enum": [
              "openai",
              "deepseek",
              "claude",
              "custom"
            ]
          },
          "ai_configs_update": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "openai",
                  "deepseek",
                  "claude",
                  "custom"
                ]
              },
              "key": {
                "type": "string",
                "nullable": true
              },
              "model": {
                "type": "string",
                "nullable": true
              },
              "baseUrl": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "string",
                    "enum": [
                      ""
                    ]
                  },
                  {}
                ]
              }
            },
            "required": [
              "provider"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get the caller's settings",
        "operationId": "getV1Settings",
        "description": "AI provider credentials are never returned — each provider reports `hasKey: boolean` only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Settings row missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update settings",
        "operationId": "patchV1Settings",
        "description": "`ai_configs_update` is the write-only lane for provider credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatchBody"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — see `error.fields` for the offending inputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled server fault. The response carries a `requestId` that appears in the server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}